implement shutdown/restart integration test
